Fixes audio in samples to work with dagonruby v241
Running the sample in the current dragonruby version 241, it crashes with the message the used `queue_sound` method on the GTK runtime is missing a parameter.
Sometime in the last year, the `queue_sound` method changed expecting a gain (float 0-1) parameter.
See https://github.com/DragonRuby/dragonruby-game-toolkit-contrib/blob/main/dragon/runtime.rb#L886

This did not impact the public interface using `arg.outputs.sounds << "path"` notation, which used in all documentation and examples.
I therefore switched the code to this notation and now the example is working fine again.